@@ -1,7 +1,7 @@
 import Swagger from "swagger-client"
 import * as configsWrapActions from "./configs-wrap-actions"
 
-module.exports = function({ configs, getConfigs }) {
+export default function({ configs, getConfigs }) {
   return {
     fn: {
       fetch: Swagger.makeHttp(configs.preFetch, configs.postFetch),
