@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 import PropTypes from "prop-types"
 import { OAS3ComponentWrapFactory } from "../helpers"
-import { Model } from "core/components/model"
+import Model from "core/components/model"
 
 class ModelComponent extends Component {
   static propTypes = {
