@@ -6,11 +6,9 @@ describe("bug #4196: HTTP basic auth credential retention", function () {
             .page.main()
 
         client.waitForElementVisible(".download-url-input:not([disabled])", 5000)
-            .pause(80)
             .clearValue(".download-url-input")
             .setValue(".download-url-input", "http://localhost:3230/test-specs/bugs/4196.yaml")
             .click("button.download-url-button")
-            .pause(1000)
 
         done()
     })
@@ -36,7 +34,7 @@ describe("bug #4196: HTTP basic auth credential retention", function () {
             .waitForElementVisible(`section>input[type="password"]`, 5000)
             .setValue(`section>input[type="password"]`, "bbb") // Set password
             .click(".auth-btn-wrapper button:nth-child(1)") // Click Authorize
-            .pause(5000)
+            .pause(1000)
             .assert.containsText("div.wrapper:nth-child(4)>code", "bbb")
 
         client.end()
